Ajustado migrações para criar carga horária como integer; carolinesalib/ants-for-education#1

+++ db/migrate/20170811230256_create_series.rb
@@ -3,7 +3,7 @@ class CreateSeries < ActiveRecord::Migration[5.0]
     create_table :series do |t|
       t.integer :ieducar_code
       t.string :name
-      t.decimal :course_load
+      t.integer :course_load
       t.integer :school_days
       t.integer :interval
